👌 IMPROVE: cleaned up lessons pages
