+++ src/components/LessonsList.svelte
@@ -1,17 +1,11 @@
 <script>
   import { createEventDispatcher } from "svelte";
   import { url } from "@sveltech/routify";
+  import { slugify } from "../utils";
   export let baseURL = "web-lessons";
   export let lessons = [];
 
   const dispatch = createEventDispatcher();
-
-  function slugify(name) {
-    return name
-      .toLowerCase()
-      .split(" ")
-      .join("_");
-  }
 
   function markComplete(id) {
     dispatch("completed", id);
