+++ src/components/common-ui/CodeBlock.svelte
@@ -1,0 +1,5 @@
+<script>
+  export let data = {};
+</script>
+
+<pre>{JSON.stringify(data, null, 2)}</pre>
