+++ src/components/index.js
@@ -3,7 +3,6 @@ import LessonsList from "./LessonsList.svelte";
 
 import AdminLessonsList from "./Admin/LessonsList.svelte";
 import AdminLessonView from "./Admin/LessonView.svelte";
-import AdminTypingLesson from "./Admin/Typing/Lesson.svelte";
 import AdminTypingSteps from "./Admin/Typing/Steps.svelte";
 
-export { Nav, LessonsList, AdminLessonsList, AdminLessonView, AdminTypingLesson, AdminTypingSteps };
+export { Nav, LessonsList, AdminLessonsList, AdminLessonView, AdminTypingSteps };
