+++ src/pages/typing-lessons/[lesson].svelte
@@ -2,34 +2,30 @@
   import { onMount } from "svelte";
   import { state, updateTypingLesson } from "../../store";
   import { goto, params } from "@sveltech/routify";
+  import { unslugify } from "../../utils";
   import TypingGUI from "../../components/TypingGUI/TypingGUI.svelte";
 
   const { typingLessons } = state;
   let specificLesson = null;
 
+  onMount(() => {
+    if (!$typingLessons.length) {
+      $goto("typing-lessons");
+    }
+    findLesson();
+  });
+
   function findLesson() {
-    const title = $params.lesson
-      .split("_")
-      .join(" ")
-      .toLowerCase();
+    const title = unslugify($params.lesson);
     specificLesson = $typingLessons.filter(
       (lesson) => lesson.title.toLowerCase() == title,
     )[0];
   }
 
-  onMount(() => {
-    if (!$typingLessons.length) {
-      $goto("/typing-lessons");
-    }
-    findLesson();
-  });
-
   function markComplete({ detail: id }) {
     updateTypingLesson(id);
-    $goto("/typing-lessons");
+    $goto("typing-lessons");
   }
 </script>
 
-<div class="container">
-  <TypingGUI lesson={specificLesson} on:sectionFinished={markComplete} />
-</div>
+<TypingGUI lesson={specificLesson} on:sectionFinished={markComplete} />
