+++ src/pages/web-lessons/[lesson].svelte
@@ -1,36 +1,31 @@
 <script>
   import { onMount } from "svelte";
   import { state, updateWebLesson } from "../../store";
-  import { url, goto, params } from "@sveltech/routify";
+  import { goto, params } from "@sveltech/routify";
+  import { unslugify } from "../../utils";
   import CodeGUI from "../../components/CodeGUI/CodeGUI.svelte";
 
   const { webLessons } = state;
-
   let specificLesson = null;
 
+  onMount(() => {
+    if (!$webLessons.length) {
+      $goto("web-lessons");
+    }
+    findLesson();
+  });
+
   function findLesson() {
-    const title = $params.lesson
-      .split("_")
-      .join(" ")
-      .toLowerCase();
+    const title = unslugify($params.lesson);
     specificLesson = $webLessons.filter(
       (lesson) => lesson.title.toLowerCase() == title,
     )[0];
   }
 
-  onMount(() => {
-    if (!$webLessons.length) {
-      $goto("/web-lessons");
-    }
-    findLesson();
-  });
-
   function markComplete({ detail: id }) {
     updateWebLesson(id);
-    $goto("/web-lessons");
+    $goto("web-lessons");
   }
 </script>
 
-<div class="container">
-  <CodeGUI lesson={specificLesson} on:sectionFinished={markComplete} />
-</div>
+<CodeGUI lesson={specificLesson} on:sectionFinished={markComplete} />
