+++ src/utils/url-helpers.js
@@ -1,0 +1,7 @@
+export const slugify = (name = "") => {
+  return name.toLowerCase().split(" ").join("_");
+};
+
+export const unslugify = (name = "") => {
+  return name.split("_").join(" ").toLowerCase();
+};
